+++ ReleaseNotes/release_notes.md
@@ -4,7 +4,10 @@
    1. add `pretty_array` function for nice print of arrays/sequences
    1. add some auxillary methods for matrices: `(min/max/minabs/maxabs)_element` and `(min/max)_diagonal` and `(min/max/minabs/maxabs)_element_index`
    1. improve printout of `SVectorwithError`
+   1. add `pos_error/neg_error/errors` properteis to `Ostap::Math::ValueWithError` object `
+   1. add column `@limit?` for printput of `RooFitResult` object to shwo the distance to the limits (if any). Distanc eof <3sigma, an d<5 sigma ar ecolored. Distances > 10sigma are omitted.
     
+    `   
 ## Backward incompatible
 
 ## Bug fixes